+++ packages/terra-alert/tests/nightwatch/AlertTitle.jsx
@@ -3,8 +3,7 @@ import Base from 'terra-base';
 import Alert from '../../lib/Alert';
 
 const locale = document.getElementsByTagName('html')[0].getAttribute('lang');
-// eslint-disable-next-line no-console
-console.log('Alerts of the standard types with custom titles');
+
 export default () =>
   <Base locale={locale}>
     <Alert id="alertAlert" type={Alert.Opts.Types.ALERT} title="Alert_Alert:" >Alert of type Alert with custom title</Alert>