+++ packages/terra-content-container/src/ContentContainer.jsx
@@ -12,6 +12,10 @@ const propTypes = {
    */
   header: PropTypes.node,
   /**
+   * The footer element to be placed within the footer area of the container.
+   */
+  footer: PropTypes.node,
+  /**
    * The children to be placed within the main content area of the container.
    */
   children: PropTypes.node,
@@ -23,12 +27,14 @@ const propTypes = {
 
 const defaultProps = {
   header: undefined,
+  footer: undefined,
   children: undefined,
   fill: false,
 };
 
 const ContentContainer = ({
   header,
+  footer,
   children,
   fill,
   ...customProps
@@ -41,14 +47,17 @@ const ContentContainer = ({
 
   return (
     <div {...customProps} className={contentLayoutClassNames}>
-      <div className={cx('header')}>
+      {header && <div className={cx('header')}>
         {header}
-      </div>
+      </div>}
       <div className={cx('main')}>
         <div className={cx('normalizer')}>
           {children}
         </div>
       </div>
+      {footer && <div className={cx('footer')}>
+        {footer}
+      </div>}
     </div>
   );
 };