+++ packages/terra-tabs/tests/nightwatch/TabPane/CustomDisplayTabPane.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import IconSearch from 'terra-icon/lib/icon/IconSearch';
+import TabPane from '../../../lib/TabPane';
+
+const CustomDisplayTabPane = () => (
+  <div role="tablist" style={{ display: 'flex' }}>
+    <TabPane
+      label="Custom Display"
+      id="customDisplay"
+      customDisplay={
+        <div id="customDisplayContent">
+          <span style={{ marginRight: '15px' }}>
+            Custom Display
+          </span>
+          <IconSearch />
+        </div>
+      }
+    />
+  </div>
+);
+
+export default CustomDisplayTabPane;