+++ packages/terra-alert/CHANGELOG.md
@@ -3,7 +3,13 @@ Changelog
 
 Unreleased
 ----------
-- Fixed lint errors
+### Fixed
+* Fixed lint errors
+* Fixed color contrast ratios to pass accessibility testing
+* Fixed lint errors
+
+### Removed
+* Removed console logs from test pages.
 
 1.13.0 - (November 28, 2017)
 ------------------